基本完成了商品增改查功能

+++ common/src/main/java/cn/edu/nju/software/common/pojo/AddressInfo.java
@@ -17,13 +17,13 @@ public class AddressInfo {
      */
     @Getter
     @Setter
-    private long longtitude;
+    private double longtitude;
     /**
      * 纬度
      */
     @Getter
     @Setter
-    private long latitude;
+    private double latitude;
     /**
      * 地址描述
      */
